+++ demo/app/demo-performance/index.js
@@ -168,7 +168,6 @@ export default class App extends Component {
         itemTouchSendsClick={false}
         stackItems
         itemHeightRatio={0.75}
-        showCursorLine
         // resizeDetector={containerResizeDetector}
 
         defaultTimeStart={defaultTimeStart}