+++ src/lib/markers/markerType.js
@@ -1,0 +1,5 @@
+export const TimelineMarkerType = {
+  Today: 'Today',
+  Custom: 'Custom',
+  Cursor: 'Cursor'
+}